pyproject.toml
[tool.poetry]
name = "etl-pipeline"
version = "0.1.0"
description = "ETL Pipeline with Prefect, AWS S3, and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
packages = [{ include = "etl_pipeline" }]

[tool.poetry.dependencies]
python = "^3.10"
prefect = "^2.10.0"
pandas = "^2.0.0"
boto3 = "^1.28.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
